byte to GB

+++ techniques.h
@@ -33,11 +33,11 @@ public:
      @param model
      @param avail_mem The available memory measured by "sizeof(double)"
      */
-    void materialize(string table_T, setting _setting, double *&model, long avail_mem);
-    void stream(string table_S, string table_R, setting _setting, double *&model, long avail_mem);
-    void factorize(string table_S, string table_R, setting _setting, double *&model, long avail_mem);
-    void materializeBCD(string table_T, setting _setting, double *&model, int block_size, long avail_mem);
-    void factorizeBCD(string table_S, string table_R, setting _setting, double *&model, int block_size, long avail_mem);
+    void materialize(string table_T, setting _setting, double *&model, double avail_mem);
+    void stream(string table_S, string table_R, setting _setting, double *&model, double avail_mem);
+    void factorize(string table_S, string table_R, setting _setting, double *&model, double avail_mem);
+    void materializeBCD(string table_T, setting _setting, double *&model, int block_size, double avail_mem);
+    void factorizeBCD(string table_S, string table_R, setting _setting, double *&model, int block_size, double avail_mem);
     bool stop(int k, double r_prev, double r_curr, setting &setting);
     void SGD(vector< vector<double> > data, setting _setting, double *&model, int feature_num);
     void BGD(vector< vector<double> > data, setting _setting, double *&model, int feature_num);
